requirements.txt
supabase>=2.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
anthropic>=0.40.0
resend>=2.0.0
schedule>=1.2.0
streamlit>=1.30.0
